--user=admin 添加上传文件功能100%/100%

+++ src/main/java/top/copying/blogs/model/dto/FileDto.java
@@ -24,7 +24,7 @@ public class FileDto implements Serializable {
     String upFileName;
 
     @ApiModelProperty("文件大小")
-    Integer fileSize;
+    Long fileSize;
 
     @ApiModelProperty("文件类型")
     String fileType;
